Support 4-Byte AS Numbers
Now neutron_dynamic_routing supports 2 byte AS numbers only.
This patch expands AS numbers constraint so that it supports
4 byte AS numbers. It expects that operators use asplain
notation to set AS numbers[1]. That's backward compatible with
existing 2 byte AS numbers.

[1]: https://tools.ietf.org/html/rfc5396

Closes-Bug: #1573092

+++ neutron_dynamic_routing/db/bgp_db.py
@@ -93,7 +93,7 @@ class BgpSpeaker(model_base.BASEV2,
     __tablename__ = 'bgp_speakers'
 
     name = sa.Column(sa.String(255), nullable=False)
-    local_as = sa.Column(sa.Integer, nullable=False, autoincrement=False)
+    local_as = sa.Column(sa.BigInteger(), nullable=False, autoincrement=False)
     advertise_floating_ip_host_routes = sa.Column(sa.Boolean, nullable=False)
     advertise_tenant_networks = sa.Column(sa.Boolean, nullable=False)
     peers = orm.relationship(BgpSpeakerPeerBinding,
@@ -118,7 +118,7 @@ class BgpPeer(model_base.BASEV2,
     name = sa.Column(sa.String(255), nullable=False)
     peer_ip = sa.Column(sa.String(64),
                         nullable=False)
-    remote_as = sa.Column(sa.Integer, nullable=False, autoincrement=False)
+    remote_as = sa.Column(sa.BigInteger(), nullable=False, autoincrement=False)
     auth_type = sa.Column(sa.String(16), nullable=False)
     password = sa.Column(sa.String(255), nullable=True)
